+++ core/components/cronmanager/model/cronmanager/cronmanager.class.php
@@ -9,7 +9,7 @@
  * @subpackage classfile
  */
 
-require_once dirname(__DIR__, 2) . '/vendor/autoload.php';
+require_once dirname(dirname(__DIR__)) . '/vendor/autoload.php';
 
 /**
  * class CronManager