+++ core/components/cronmanager/src/CronManager.php
@@ -40,7 +40,7 @@ class CronManager
      * The version
      * @var string $version
      */
-    public $version = '1.3.1';
+    public $version = '1.3.2';
 
     /**
      * The class options