+++ core/components/cronmanager/docs/changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] - 2024-08-12
+
+### Changed
+
+- Calculate the nextrun on base of the last run and not on the execution time of the cron connector
+- Add an editable nextrun field in the cronjob create/update window
+
 ## [1.3.1] - 2024-02-14
 
 ### Fixed